zjzrequirements.txt
absl-py==2.1.0
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
blinker==1.7.0
cachetools==5.5.2
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==45.0.5
cycler==0.12.0
deepface==0.0.93
defusedxml==0.7.1
face-recognition==1.2.3
face_recognition_models==0.3.0
filelock==3.13.3
fire==0.7.0
flasgger==0.9.7.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2024.3.1
gast==0.4.0
gdown==5.2.0
google-auth==2.40.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.62.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.10.0
humanfriendly==10.0
idna==3.6
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lapx==0.5.11.post1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mistune==3.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
mtcnn==0.1.1
mysql-connector-python==9.3.0
namex==0.0.7
networkx==3.2.1
numpy ==1.24.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nvjitlink-cu12==12.9.86
nvidia-pyindex==1.0.9
oauthlib==3.3.1
onnxruntime==1.17.1
onnxruntime-gpu==1.17.1
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.1
pillow==10.3.0
Pipeless==2.0.0
protobuf==3.19.6
psutil==5.9.8
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybboxes==0.1.6
pycparser==2.22
Pygments==2.17.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.1.2
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.7.1
rpds-py==0.26.0
rsa==4.9.1
sahi==0.11.30
scipy==1.12.0
seaborn==0.13.2
sentencepiece==0.2.0
shapely==2.0.7
simple-websocket==1.1.0
six==1.16.0
soupsieve==2.7
SQLAlchemy==2.0.41
supervision==0.25.1
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvision==0.17.2+cu118
tqdm==4.66.2
typing_extensions==4.14.1
tzdata==2024.1
ultralytics==8.1.41
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.14.1
wsproto==1.2.0
zipp==3.23.0
